pyproject.toml
[project]
name = "fire-crust-back-end"
version = "0.1.0"
description = "SaaS backend for Fire Crust pizza ordering built with FastAPI and PostgreSQL"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "sqlalchemy>=2.0.36,<3.0.0",
  "psycopg[binary]>=3.2.0,<4.0.0",
  "pydantic-settings>=2.6.0,<3.0.0",
  "python-jose[cryptography]>=3.3.0,<4.0.0",
  "passlib[argon2]>=1.7.4,<2.0.0",
  "argon2-cffi>=21.3.0,<30.0.0",
  "python-multipart>=0.0.12,<1.0.0",
  "email-validator>=2.2.0,<3.0.0",
  "python-dotenv>=1.0.1,<2.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
  "httpx>=0.27.0,<1.0.0",
  "pytest-asyncio>=0.24.0,<1.0.0",
  "aiosqlite>=0.20.0,<1.0.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
